+++ Zotero/Controllers/API/Responses/AuthorizeUploadResponse.swift
