+++ Zotero/Controllers/API/Responses/CollectionResponse.swift
