+++ Zotero/Controllers/API/Responses/DeletionsResponse.swift
