+++ Zotero/Controllers/API/Responses/GroupResponse.swift
