+++ Zotero/Controllers/API/Responses/LinksResponse.swift
