+++ Zotero/Controllers/API/Responses/SchemaResponse.swift
