+++ Zotero/Controllers/API/Responses/SearchesResponse.swift
