+++ Zotero/Controllers/API/Responses/SettingsResponse.swift
